The use of AI_NUMERICHOST in init_transport() is broken.
In its current form, when seeing a dotted quad or a numeric IPv6 address,
it will clear all flags in hints.ai_flags (which has been set to AI_PASSIVE
a few lines above).

What this code should really be doing is to *set* AI_NUMERICHOST if and only
if it sees a numeric address.

Signed-off-by: Olaf Kirch <okir@suse.de>