rpcinfo: introduce new helper function ip_getclient
This function tries to obtain the address of a given host/program/version,
using the specified protocol (one of udp or tcp), and create a client
handle for it.

This loops over all netconfig entries (according to the order given
by netpath and the config file), and tries to resolve the hostname,
and obtain the address using rpcb_getaddr.

Signed-off-by: Olaf Kirch <okir@suse.de>