rpcinfo: make -t/-u options support IPv6 addresses, too
Currently, rpcinfo supports two ways to test whether a given
service is alive.

For IPv4 hosts, you can use -u and -t. But for IPv6, you need
to use "rpcinfo -a" and specify the host through a universal
address, which is fairly inconvenient.

This patch modifies ip_ping to allow using -u and -t for IPv6
addresses and hosts as well.

Signed-off-by: Olaf Kirch <okir@suse.de>